do not read the entry when cursor option used
if the user includes a cursor in the request, we should not read the same
entry, but move on to the next one. A good example is a server sent event
field "id". If the connection drops the browser will reconnect providing the
`Last-Event-ID` header to the server. Since browser is aware of the last cursor
it means it has the very last entry and we have to provide all the following
after the cursor.