handle journal files rotation properly
add a new journal reader function `Follow()` which takes a timeout interval
and a io.Writer. The function takes care of tracking if the journald files
were rotated due to normal journald rotation policy. In this case, it will
close the rotated file descriptiors and reopened them with original user
parameters.