refactor logging api v2
- refactor GET parameters handlers:
  - add a parameter handler for each GET parameter.
  - add a function `buildOpts` which is responsible for building the reader
    parameters based on GET request parameters.
- fix a bug, which should ignore the initial request parameters if
  `Last-Event-ID` request header is used by a client.

Problem: suppose a client library makes a request to get 10 last lines from
then end of the log. The request would look like: '/?cursor=END&skip=-10'.
That will work fine, however if a connection timeout occurs, the client is
responsible to reconnect and pass `Last-Event-ID` in request headers to start
consuming the logs where the client dropped off. So the server should give back
all logs after the supplied cursor position. However the reconnect request will
also include the initial parameters '/?cursor=END&skip=-10'. This will tell the
server to ignore the `Last-Event-ID` cursor position and instead read the 10
last entries again. The fix here would be to ignore the other GET parameters if
`Last-Event-ID` is used by a client in the request.