Fixed file streaming using the Mesos File API.
With this commit we address the problems in DCOS-41248.
In order to not overload the Mesos agent with GET requests, we will now
wait one second if we are reading 0 bytes from the Mesos file.

This fix is temporary and not ideal but feels responsive enough to the
enduser and is a better alternative to what we have today.
The right fix here should be a change to the Read semantics so that a
read will block until there is data available.

https://jira.mesosphere.com/browse/DCOS-41248